client: Move extend api methods into each module
