+++ iota-client/extended/mod.rs
@@ -1,5 +1,8 @@
+pub(crate) mod get_bundle;
 pub(crate) mod get_inputs;
 pub(crate) mod get_new_address;
 pub(crate) mod prepare_transfers;
 pub(crate) mod send_transfers;
 pub(crate) mod send_trytes;
+pub(crate) mod store_and_broadcast;
+pub(crate) mod traverse_bundle;
