+++ iota-client/extended/send_transfers.rs
@@ -1,4 +1,8 @@
-use iota_model::Inputs;
+use iota_model::{Inputs, Transaction, Transfer};
+
+use crate::client::Client;
+use crate::options::{PrepareTransfersOptions, SendTrytesOptions};
+use crate::Result;
 
 /// SendTransferOptions
 #[derive(Clone, Debug, PartialEq)]
@@ -36,5 +40,46 @@ impl<'a, 'b, 'c> Default for SendTransferOptions<'a, 'b, 'c> {
             security: 3,
             hmac_key: None,
         }
+    }
+}
+
+impl<'a> Client<'a> {
+    /// Prepares and sends a slice of transfers
+    /// This helper does everything for you, PoW and such
+    ///
+    /// * `transfers` - A slice of transfers to send
+    /// * `seed` - The wallet seed to use
+    /// * `depth` - The depth to search when looking for transactions to approve
+    /// * `min_weight_magnitude` - The PoW difficulty factor (14 on mainnet, 9 on testnet)
+    /// * `local_pow` - Whether or not to do local PoW
+    /// * `options` - See `SendTransferOptions`
+    pub fn send_transfers(
+        &mut self,
+        transfers: impl Into<Vec<Transfer>>,
+        seed: &str,
+        options: SendTransferOptions<'_, '_, '_>,
+    ) -> Result<Vec<Transaction>> {
+        let transfers = transfers.into();
+        let trytes = self.prepare_transfers(
+            seed,
+            transfers,
+            PrepareTransfersOptions {
+                inputs: options.inputs,
+                remainder_address: options.remainder_address,
+                security: options.security,
+                hmac_key: options.hmac_key,
+            },
+        )?;
+        let t = self.send_trytes(
+            &trytes,
+            SendTrytesOptions {
+                depth: options.depth,
+                min_weight_magnitude: options.min_weight_magnitude,
+                local_pow: options.local_pow,
+                threads: options.threads,
+                reference: options.reference,
+            },
+        )?;
+        Ok(t)
     }
 }
