+++ src/prelude.rs
@@ -1,6 +1,6 @@
 pub use crate::{
-    iota_client, iota_constants, iota_conversion, iota_crypto, iota_model, iota_pow,
-    iota_signing, iota_utils, iota_validation,
+    iota_client, iota_constants, iota_conversion, iota_crypto, iota_model, iota_pow, iota_signing,
+    iota_utils, iota_validation,
 };
 /// This trait provides conversion methods for various types into trinary
 pub use iota_conversion::Trinary;
